requirements.txt
annotated-types==0.6.0
anyio==4.2.0
beanie==1.24.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
et-xmlfile==1.1.0
fastapi==0.109.0
h11==0.14.0
httptools==0.6.1
idna==3.6
lazy-model==0.2.0
motor==3.3.2
numpy==1.26.3
openpyxl==3.1.2
pandas==2.1.4
pydantic==2.5.3
pydantic_core==2.14.6
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.0
starlette==0.35.1
toml==0.10.2
typing_extensions==4.9.0
tzdata==2023.4
uvicorn==0.26.0
watchfiles==0.21.0
websockets==12.0
